backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-multipart
passlib[bcrypt]
pyjwt
langchain-groq
langchain-community
langchain-text-splitters
sentence-transformers
chromadb
pypdf
aiofiles
